fix: fix displaying card
