+++ src/components/characterCard/index.ts
@@ -1,0 +1,2 @@
+import CharacterCard from './CharacterCard';
+export { CharacterCard };
