+++ src/components/detailedSection/cardWrapper/index.ts
@@ -1,0 +1,1 @@
+export { CardWrapper } from './CardWrapper';
