+++ src/types/types.ts
@@ -1,4 +1,4 @@
-import { ReactNode } from 'react';
+import { ReactNode, JSX } from 'react';
 
 export interface ICharacterCardProps {
   character: IStarWarsCharacter;
